[API-33] default domain for enable CIFS service
If no domain name is supplied when enables CIFS service on NAS server, it
will check the DNS server available and use the domain name of the DNS
server as the default domain.

[API-34] Add SMB name already used exception

If SMB name is already used when enables CIFS service,
`UnityXmbNameInUseError` will be thrown.

[API-35] NAS server as CIFS server in create share

Since we could hide the CIFS server concept in API.
It's also OK for user to pass in the CIFS server instance in create CIFS
share method.

Modify the `get` class method of `UnityCifsServer` to accept instance of
`UnityNasServer`.  It will return the first CIFS server of that NAS
server.

Pop the version to 0.1.5