[API-47] Extra error log for heartbeat
An error log is written for each heartbeat call.
Root cause is that we check error for the heartbeat call.

Solution is to remove the error log in `raise_if_err` function.
The principle is that logs are not written when it's thrown.

And add a function to check specified error.

The exception catcher will have the right to decide whether to write
log or not.

Remove the time out error because it's included in
`VNXSpNotAvailableError`.