[API-51] get_sg doesn't return the host info
The property of storage group doesn't contains the HBA/SP pair.

The root cause is "HBA/SP Pairs:" is not properly parsed.
In the parse configuration we assume that "HBA/SP Pairs" has a peer
property named "HLU/ALU Pairs".  This peer property does not exists if the
storage group doesn't have any HLU mapping.

Modify the end pattern of this property to accept both "HLU/ALU Pairs" and
"Shareable".  And the "HBA/SP Pairs" are property extracted and parsed.

Other enhancements are utilize the `JsonPrinter` to print the string
representation of `VNXHbaPort` and `VNXStorageGroupHBA`.