[API-85] Default VNX list property to empty.
For VNX resources, if the property is a list and is
not present, default the value to an empty tuple.

This change does not affect Unity resource.
For unity resource, the output property strictly
align with the output of REST API.  The reason is
an empty list has a different meaning to a null or
not presented value in REST API.