[API-104] Delete non-existed LUN raises ValueError
Deleting a non-existed LUN with `force` option will raise an unexpected
ValueError.

The root cause is we tried to detach the not-existed LUN first before
deleting it.

The fix is to catch ValueError in detach from storage group or remove from
consistency group.  If LUN id is not valid, skip this LUN for the
operation.