[API-110] remove FS failed.
Remove file system in Manila failed because it checks the available
snapshots of the system first.

And when a snapshot is just deleted, we could still see snapshots as a
child of the filesystem.

The fix is add another `has_snap` function which returns `False` if no
shapshots are available for the specified filesystem or all of them are in
`destroying` state.