Fix alu/hlu cache issue
Under certain circumstances, the alu/hlu cache was flushed by an
sg.update(), thus the detach of the alu was skipped as not found
in the cache.

An unbind LUN error will be thrown when deleting the LUN:

 Cannot unbind LUN because it's contained in a Storage Group

This commit will update the cache incrementally when calling
sg.update, addtionally more logs are added around the cache for
later triage.